query_eval: split the eval dispatcher into per-node-type modules
`query_eval/src/eval.rs` had grown to 1050 lines holding three unrelated
concerns: the evaluator infrastructure (`Config`, `Evaluated`), the AST
dispatcher, and twelve independent per-node-type evaluators.

Move each evaluator into its own module, mirroring the per-node-type layout
this crate already uses for its integration tests, so each evaluator sits next
to a test file of the same name. `lib.rs` keeps the shared parts: `Config`,
`Evaluated`, the `eval_node` dispatcher and its `qast_iterate` entry point, and
the helpers for evaluating a child node and for delegating an unported node
to C.

The modules go at the crate root rather than under an `eval` module. The crate
has a single concern, so the extra level bought nothing but a stutter in
`query_eval::eval::eval_node`, and left `lib.rs` a pass-through that only
re-exported `eval`. The in-flight `QN_TAG` port already places its evaluator at
the crate root, so this also makes that module consistent with the rest rather
than an exception.

This renames the public paths — `query_eval::eval::eval_node` becomes
`query_eval::eval_node`, and likewise for `qast_iterate`, `Config`,
`Evaluated` and `EvalResult`. `query_eval_ffi` and the integration tests are
updated accordingly. No logic, doc-comment or SAFETY-comment text is edited:
every line of the old `eval.rs` is present unchanged, only redistributed.
