+++ src/redisearch_rs/query_eval/tests/integration/ids.rs
@@ -9,7 +9,7 @@
 
 //! QN_IDS → IdList
 
-use query_eval::{QueryEvalContext, QueryNodeMut, eval, eval::Config};
+use query_eval::{Config, QueryEvalContext, QueryNodeMut, eval_node};
 use query_types::QueryNodeType;
 use rqe_iterators::{IteratorType, RQEIterator};
 
@@ -28,7 +28,7 @@ fn eval_ids_with_pre_resolved_doc_ids() {
     mock_node.set_ids(keys.as_ptr(), doc_ids.as_mut_ptr(), keys.len());
     let node = unsafe { QueryNodeMut::new(mock_node.as_non_null()) };
 
-    let mut it = eval::eval_node(&mut ctx, node, Config::default())
+    let mut it = eval_node(&mut ctx, node, Config::default())
         .expect("should not be None")
         .into_boxed();
 
@@ -58,7 +58,7 @@ fn eval_ids_deduplicates() {
     mock_node.set_ids(keys.as_ptr(), doc_ids.as_mut_ptr(), keys.len());
     let node = unsafe { QueryNodeMut::new(mock_node.as_non_null()) };
 
-    let mut it = eval::eval_node(&mut ctx, node, Config::default())
+    let mut it = eval_node(&mut ctx, node, Config::default())
         .expect("should not be None")
         .into_boxed();
 
@@ -82,7 +82,7 @@ fn eval_ids_filters_zero_doc_ids() {
     mock_node.set_ids(keys.as_ptr(), doc_ids.as_mut_ptr(), keys.len());
     let node = unsafe { QueryNodeMut::new(mock_node.as_non_null()) };
 
-    let mut it = eval::eval_node(&mut ctx, node, Config::default())
+    let mut it = eval_node(&mut ctx, node, Config::default())
         .expect("should not be None")
         .into_boxed();
 
@@ -104,7 +104,7 @@ fn eval_ids_all_zero_produces_empty_list() {
     mock_node.set_ids(keys.as_ptr(), doc_ids.as_mut_ptr(), keys.len());
     let node = unsafe { QueryNodeMut::new(mock_node.as_non_null()) };
 
-    let mut it = eval::eval_node(&mut ctx, node, Config::default())
+    let mut it = eval_node(&mut ctx, node, Config::default())
         .expect("should not be None")
         .into_boxed();
 
@@ -121,7 +121,7 @@ fn eval_ids_empty_keys() {
     mock_node.set_ids(std::ptr::null(), std::ptr::null_mut(), 0);
     let node = unsafe { QueryNodeMut::new(mock_node.as_non_null()) };
 
-    let mut it = eval::eval_node(&mut ctx, node, Config::default())
+    let mut it = eval_node(&mut ctx, node, Config::default())
         .expect("should not be None")
         .into_boxed();
 
@@ -168,7 +168,7 @@ mod ids_doctable {
         mock_node.set_ids(keys.as_ptr(), std::ptr::null_mut(), keys.len());
         let node = unsafe { QueryNodeMut::new(mock_node.as_non_null()) };
 
-        let mut it = eval::eval_node(&mut ctx, node, Config::default())
+        let mut it = eval_node(&mut ctx, node, Config::default())
             .expect("should not be None")
             .into_boxed();
 
@@ -201,7 +201,7 @@ mod ids_doctable {
         mock_node.set_ids(keys.as_ptr(), std::ptr::null_mut(), keys.len());
         let node = unsafe { QueryNodeMut::new(mock_node.as_non_null()) };
 
-        let mut it = eval::eval_node(&mut ctx, node, Config::default())
+        let mut it = eval_node(&mut ctx, node, Config::default())
             .expect("should not be None")
             .into_boxed();
