+++ src/redisearch_rs/query_eval/tests/integration/numeric.rs
@@ -18,10 +18,7 @@
 use index_result::RSIndexResult;
 use inverted_index::NumericFilter;
 use query::mock::MockQueryNode;
-use query_eval::{
-    QueryEvalContext, QueryNodeMut,
-    eval::{self, Config, EvalResult},
-};
+use query_eval::{Config, EvalResult, QueryEvalContext, QueryNodeMut, eval_node};
 use query_types::QueryNodeType;
 use rqe_iterators::{IteratorType, IteratorsConfig, RQEIterator};
 use rqe_iterators_test_utils::{GlobalGuard, TestContext};
@@ -99,7 +96,7 @@ impl NumericFixture {
     fn eval(&mut self) -> Option<EvalResult<'_>> {
         // SAFETY: `self.node` is a valid, live `RSQueryNode` for the call.
         let node_ref = unsafe { QueryNodeMut::new(self.node.as_non_null()) };
-        eval::eval_node(&mut self.ctx, node_ref, Config::default()).map(|e| e.into_boxed())
+        eval_node(&mut self.ctx, node_ref, Config::default()).map(|e| e.into_boxed())
     }
 }
