+++ src/redisearch_rs/query_eval/tests/integration/optional.rs
@@ -9,7 +9,7 @@
 
 //! QN_OPTIONAL → Optional / Wildcard (via the reducer shortcircuits)
 
-use query_eval::{QueryEvalContext, QueryNodeMut, eval, eval::Config};
+use query_eval::{Config, QueryEvalContext, QueryNodeMut, eval_node};
 use query_types::QueryNodeType;
 use rqe_iterators::{IteratorType, RQEIterator};
 
@@ -30,7 +30,7 @@ fn eval_optional_empty_child_falls_back_to_wildcard() {
     opt.set_children(&[null_child.as_ptr()]);
     let node = unsafe { QueryNodeMut::new(opt.as_non_null()) };
 
-    let mut it = eval::eval_node(&mut ctx, node, Config::default())
+    let mut it = eval_node(&mut ctx, node, Config::default())
         .expect("should not be None")
         .into_boxed();
 
@@ -59,7 +59,7 @@ fn eval_optional_wildcard_child_passes_through() {
     opt.set_children(&[wc_child.as_ptr()]);
     let node = unsafe { QueryNodeMut::new(opt.as_non_null()) };
 
-    let mut it = eval::eval_node(&mut ctx, node, Config::default())
+    let mut it = eval_node(&mut ctx, node, Config::default())
         .expect("should not be None")
         .into_boxed();
 
@@ -114,7 +114,7 @@ mod optional {
         opt.set_children(&[ids_child.as_ptr()]);
         let node = unsafe { QueryNodeMut::new(opt.as_non_null()) };
 
-        let mut it = eval::eval_node(&mut ctx, node, Config::default())
+        let mut it = eval_node(&mut ctx, node, Config::default())
             .expect("should not be None")
             .into_boxed();
 
@@ -162,7 +162,7 @@ mod optional {
         opt.set_children(&[missing_child.as_ptr()]);
         let node = unsafe { QueryNodeMut::new(opt.as_non_null()) };
 
-        let mut it = eval::eval_node(&mut ctx, node, Config::default())
+        let mut it = eval_node(&mut ctx, node, Config::default())
             .expect("an optional node always yields an iterator")
             .into_boxed();
 
@@ -205,7 +205,7 @@ mod optional {
         opt.set_children(&[ids_child.as_ptr()]);
         let node = unsafe { QueryNodeMut::new(opt.as_non_null()) };
 
-        let mut it = eval::eval_node(&mut ctx, node, Config::default())
+        let mut it = eval_node(&mut ctx, node, Config::default())
             .expect("an optional node always yields an iterator")
             .into_boxed();
