+++ src/redisearch_rs/query_eval/tests/integration/qast_iterate.rs
@@ -7,7 +7,7 @@
  * GNU Affero General Public License v3 (AGPLv3).
 */
 
-use query_eval::{QueryEvalContext, QueryNodeMut, eval, eval::Config};
+use query_eval::{Config, QueryEvalContext, QueryNodeMut, qast_iterate};
 use query_types::QueryNodeType;
 use rqe_iterators::{IteratorType, RQEIterator};
 
@@ -20,7 +20,7 @@ fn qast_iterate_evaluates_root_node() {
     let mock_node = MockQueryNode::new(QueryNodeType::Null);
     let node = unsafe { QueryNodeMut::new(mock_node.as_non_null()) };
 
-    let mut it = eval::qast_iterate(&mut ctx, node, Config::default()).into_boxed();
+    let mut it = qast_iterate(&mut ctx, node, Config::default()).into_boxed();
 
     assert_eq!(it.type_(), IteratorType::Empty);
     assert!(it.at_eof());
