+++ src/redisearch_rs/query_eval/tests/integration/token.rs
@@ -25,10 +25,7 @@ use ffi::{
 };
 use index_result::{RSIndexResult, RSOffsetSlice};
 use query::mock::MockQueryNode;
-use query_eval::{
-    QueryEvalContext, QueryNodeMut,
-    eval::{self, Config, EvalResult},
-};
+use query_eval::{Config, EvalResult, QueryEvalContext, QueryNodeMut, eval_node};
 use query_term::RSQueryTerm;
 use query_types::QueryNodeType;
 use rqe_core::{FieldMask, RS_FIELDMASK_ALL};
@@ -115,7 +112,7 @@ impl TokenFixture {
     fn eval(&mut self) -> Option<EvalResult<'_>> {
         // SAFETY: `self.node` is a valid, live `RSQueryNode` for the call.
         let node_ref = unsafe { QueryNodeMut::new(self.node.as_non_null()) };
-        eval::eval_node(&mut self.ctx, node_ref, Config::default()).map(|e| e.into_boxed())
+        eval_node(&mut self.ctx, node_ref, Config::default()).map(|e| e.into_boxed())
     }
 }
