+++ src/redisearch_rs/query_eval/tests/integration/union.rs
@@ -9,7 +9,7 @@
 
 //! QN_UNION → Union
 
-use query_eval::{QueryEvalContext, QueryNodeMut, eval, eval::Config};
+use query_eval::{Config, QueryEvalContext, QueryNodeMut, eval_node};
 use query_types::QueryNodeType;
 use rqe_iterators::{IteratorType, RQEIterator};
 
@@ -32,7 +32,7 @@ fn eval_union_merges_children() {
     union.set_children(&[c1.as_ptr(), c2.as_ptr()]);
     let node = unsafe { QueryNodeMut::new(union.as_non_null()) };
 
-    let mut it = eval::eval_node(&mut ctx, node, Config::default())
+    let mut it = eval_node(&mut ctx, node, Config::default())
         .expect("should not be None")
         .into_boxed();
 
@@ -65,7 +65,7 @@ fn eval_union_zero_weight_takes_quick_exit() {
     union.set_children(&[c1.as_ptr(), c2.as_ptr()]);
     let node = unsafe { QueryNodeMut::new(union.as_non_null()) };
 
-    let mut it = eval::eval_node(&mut ctx, node, Config::default())
+    let mut it = eval_node(&mut ctx, node, Config::default())
         .expect("should not be None")
         .into_boxed();
 
@@ -101,7 +101,7 @@ fn eval_union_in_not_subtree_takes_quick_exit() {
     union.set_children(&[c1.as_ptr(), c2.as_ptr()]);
     let node = unsafe { QueryNodeMut::new(union.as_non_null()) };
 
-    let mut it = eval::eval_node(&mut ctx, node, Config::default())
+    let mut it = eval_node(&mut ctx, node, Config::default())
         .expect("should not be None")
         .into_boxed();
 
@@ -175,7 +175,7 @@ mod union {
         union.set_children(&[c1.as_ptr(), c2.as_ptr()]);
         let node = unsafe { QueryNodeMut::new(union.as_non_null()) };
 
-        let mut it = eval::eval_node(&mut ctx, node, Config::default())
+        let mut it = eval_node(&mut ctx, node, Config::default())
             .expect("should not be None")
             .into_boxed();
 
@@ -233,7 +233,7 @@ mod union {
         union.set_children(&[missing_child.as_ptr(), ids_child.as_ptr()]);
         let node = unsafe { QueryNodeMut::new(union.as_non_null()) };
 
-        let mut it = eval::eval_node(&mut ctx, node, Config::default())
+        let mut it = eval_node(&mut ctx, node, Config::default())
             .expect("a multi-child union always yields an iterator")
             .into_boxed();
